+++ changelog.md
@@ -1,3 +1,7 @@
+### Wireless Chargers 1.0.10
+- Fixed energy transfer when blocks report a side as being able to receive energy but which doesn't actually accept energy
+- Added Turkish translations (thanks to RuyaSavascisi!)
+
 ### Wireless Chargers 1.0.9
 - Initial release of Wireless Chargers for Forge 1.21.2 & 1.21.3
 Update changelog